chore(apollo-react): move NodeIO story
